# Docs API: Shorter string representation of types, faster code, and other smaller changes