--- a/src/MudBlazor.Docs/Components/DocsApi.razor
+++ b/src/MudBlazor.Docs/Components/DocsApi.razor
@@ -40,7 +40,7 @@
                                     <MudText Typo="Typo.h6">@($"Inherited from {((Type)context.Key).GetTypeDisplayName()}")</MudText>
                                 </MudTh>
                             }
-                            else if (_propertiesGrouping == Grouping.Category)
+                            else if (_propertiesGrouping == Grouping.Categories)
                             {
                                 <MudTh colspan="4">
                                     <MudText Typo="Typo.h6">@context.Key</MudText>
@@ -210,7 +210,7 @@
 
         foreach (var info in Type.GetPropertyInfosWithAttribute<ParameterAttribute>()
                                  .OrderBy(x => _propertiesGrouping switch {
-                                                   Grouping.Category => x.GetCustomAttribute<CategoryAttribute>()?.Order ?? int.MaxValue - 1,
+                                                   Grouping.Categories => x.GetCustomAttribute<CategoryAttribute>()?.Order ?? int.MaxValue - 1,
                                                    Grouping.Inheritance => -NumberOfAncestorClasses(BaseDefinitionClass(x)),
                                                    _ => 0
                                                })
@@ -340,13 +340,13 @@
 
     #region Grouping properties
 
-    private enum Grouping { Category, Inheritance, None }
+    private enum Grouping { Categories, Inheritance, None }
 
     private Grouping _propertiesGrouping = Grouping.Inheritance;
 
     private TableGroupDefinition<ApiProperty> PropertiesGroupDefinition => _propertiesGrouping switch
     {
-        Grouping.Category => new() { Selector = (p) => p.PropertyInfo.GetCustomAttribute<CategoryAttribute>()?.Name ?? "Misc" },
+        Grouping.Categories => new() { Selector = (p) => p.PropertyInfo.GetCustomAttribute<CategoryAttribute>()?.Name ?? "Misc" },
         Grouping.Inheritance => new() { Selector = (p) => BaseDefinitionClass(p.PropertyInfo) },
         _ => null
     };