--- a/src/MudBlazor.Docs/Components/DocsTypeInfo.razor
+++ b/src/MudBlazor.Docs/Components/DocsTypeInfo.razor
@@ -1,13 +1,13 @@
-﻿@if (NonnullableType.IsEnum)
+﻿@if (_nonnullableType.IsEnum)
 {
     <MudTooltip>
         <ChildContent>@Type.ConvertToCSharpSource()</ChildContent>
         <TooltipContent>
             enumeration type
             <MudText Align="Align.Left" Typo="Typo.body2">
-                @foreach (var name in Enum.GetNames(NonnullableType))
+                @foreach (var name in Enum.GetNames(_nonnullableType))
                 {
-                    @(NonnullableType.Name + "." + name)<br />
+                    @(_nonnullableType.Name + "." + name)<br />
                 }
             </MudText>
         </TooltipContent>
@@ -21,5 +21,10 @@ else
 @code {
     [Parameter] public Type Type { get; set; }
 
-    private Type NonnullableType => Nullable.GetUnderlyingType(Type) ?? Type;
+    private Type _nonnullableType;
+
+    protected override void OnInitialized()
+    {
+        _nonnullableType = Nullable.GetUnderlyingType(Type) ?? Type;
+    }
 }