--- a/src/MudBlazor.Docs/Components/EnumSwitch.razor
+++ b/src/MudBlazor.Docs/Components/EnumSwitch.razor
@@ -26,7 +26,5 @@
 
     [Parameter] public EventCallback<T> ValueChanged { get; set; }
 
-    [Parameter] public Color Color { get; set; }
-
     private Type Type => typeof(T);
 }