… src/MudBlazor.Docs/Extensions/TypeNameHelper.cs
… src/MudBlazor.UnitTests/Other/CategoryAttributeTests.cs