## Description

### Changed string representation of types from .NET format to C# format
Examples:   `Int32` changed to `int`, `Nullable<DateTime>` changed to `DateTime?`, and `Converter<Nullable<DateTime>, String>` changed to `Converter<DateTime?, string>`. More info: https://github.com/MudBlazor/MudBlazor/discussions/3520.

The code uses https://github.com/benaadams/Ben.Demystifier/blob/main/src/Ben.Demystifier/TypeNameHelper.cs (Apache License). The speed of new and old method is the same. I tested automatically each page, and then I tested manually the differences: all changes are correct except of type of `MudDialogInstance.Close(T returnValue)` method parameter, but it isn't important.

### The code is two times faster now
 - The `Properties`, `EventCallbacks`, and `Methods` properties were called respectively: 4, 6, and 6 times per one page, but should be only once which will speeds up the code **about two times**. Generally, C# properties are too easy to call, so I changed them to these methods: `getProperties()`, `getEventCallbacks()`, `getMethods()`. Each of these methods is called only once and the result is stored in a list. I wanted to call them in the `OnParametersSet()` method, but I doesn't work because the properties table depends not only on the `Type` parameter but also on the value of the "CATEGORIES - INHERITANCE - NONE" grouping switch. I don't know Blazor well, so I ended with calling these methods in the Razor markup code.
 - The `DocStrings.GetMemberDescription(type, member)` method is about 10 ms faster because now the `GetSaveTypename()` method is run only once per table, and not for every table row.
 - The `DocsTypeInfo` component is about 10 ms faster because now the `_nonnullableType` is computed only once in the `OnInitialized()` method.

All 3 changes speed up MudAutocomplete API page from 290 ms to 135 ms on my computer. The properties grouping switch was set to None during the measurement, but grouping is fast. I measured speed using the "Rendered in ... ms" counter.

### Other changes
 - By default, properties are grouped by a category (temporarily until it will be set in user settings).
 - I changed "CATEGORY" label to "CATEGORIES" in the "CATEGORY - INHERITANCE - NONE" switch.
 - New line characters in a text copied to clipboard make more sense now.
 - I refactored (simplified) some code.

## How Has This Been Tested?

I tested each page automatically using the https://github.com/MudBlazor/MudBlazor/discussions/3595 code, and additionally performed manual tests.

## Types of changes
<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

<!-- If you made any visual changes, provide screenshots of before/after, it its moving parts, please provide high quality gif, wemb or mp4 -->

## Checklist:
<!-- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!-- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] The PR is submitted to the correct branch (`dev`).
- [x] My code follows the code style of this project.
- [ ] I've added relevant tests.