@mikes-gh 2022-01-11
@iwis  Great work on the perf. 

Regarding using the type aliases its common for Documentation to use the Net Type Names instead of aliases.

e.g https://docs.microsoft.com/en-us/dotnet/api/system.filestyleuriparser?view=net-6.0
@MudBlazor/core-team

Regarding the switch I think of it as being being prefixed  with the word "By".

By Inheritence By Category By Class

But I am not precious about this change
@MudBlazor/core-team

In any case great work 🙏