@henon 2022-01-11
> Regarding using the type aliases its common for Documentation to use the Net Type Names instead of aliases.

Sure, but in the end it is a question of taste and how it looks. I think we had this discussion already and decided to use the aliases.