@mikes-gh 2022-01-11
> > Regarding using the type aliases its common for Documentation to use the Net Type Names instead of aliases.
> > e.g https://docs.microsoft.com/en-us/dotnet/api/system.filestyleuriparser?view=net-6.0 @MudBlazor/core-team
> 
> The .NET documentation uses also aliases and `?` instead of `Nullable` for properties (See [SortedSet.Count](https://docs.microsoft.com/en-us/dotnet/api/system.collections.generic.sortedset-1.count) and [SortedSet.Max](https://docs.microsoft.com/en-us/dotnet/api/system.collections.generic.sortedset-1.max) detail pages. It doesn't display types for properties at overview page at all - see [SortedSet](https://docs.microsoft.com/en-us/dotnet/api/system.collections.generic.sortedset-1?view=net-6.0#properties)).
> 
> MudBlazor already uses aliases for methods (see https://mudblazor.com/api/colorpicker#methods), although inconsistently because it doesn't cope with more difficult cases (see https://mudblazor.com/api/messagebox#methods). When it comes to methods, I don't think we should follow the .NET documentation overview pages because they are incomplete: properties don't have types, methods don't have result type and parameter names. The .NET documentation overview pages have always been hard to read for me - I even wrote a script that changes them: https://github.com/iwis/NET-Framework-documentation-signatures.
> 
> > Regarding the switch I think of it as being being prefixed with the word "By".
> > By Inheritence By Category By Class
> > But I am not precious about this change @MudBlazor/core-team
> 
> Currently, in the new doc v5 it is: CATEGORY | INHERITANCE | NONE on https://dev.mudblazor.com/ without the "group by:" label before the switch:
> 
> ![image](https://user-images.githubusercontent.com/8080496/149020582-3e62940f-0386-41ac-8a57-4c40434b58e1.png)
> 
> I am not sure if it is clear now, that is why I proposed: CATEGORIES | INHERITANCE | NONE:
> 
> ![image](https://user-images.githubusercontent.com/8080496/149020616-0fe26b15-6e01-440f-92d5-98c66b681649.png)
> 
> What do you think is the best? The "By Inheritence By Category By Class" isn't correct because the label of the 3rd button is "None" now - no grouping. Pay attention that on mobile we have only 360 pixels, so for longer labels we will probably need to use MudMenu, MudSelect, or MudRadioGroup.
> 
> ![image](https://user-images.githubusercontent.com/8080496/149020660-6f5b9846-a401-4fe0-8727-1c4dbad06e31.png)
I dont mind.  I think you thought about it so I say go for your feeling.

@henon. For me this is good to go.